[REFACTOR] review 반영
