+++ src/component/Nav/components/Modal.scss
@@ -12,13 +12,12 @@
     align-items: center;
 
     .modalTextContainer {
-      margin-left: 230px;
-      margin-top: 50px;
-      margin-bottom: 80px;
+      margin: 50px 0 230px 80px;
 
       .modalText {
         margin-top: 30px;
       }
+
       .linkText {
         text-decoration: none;
       }
