+++ src/pages/Main/Main.js
@@ -2,7 +2,165 @@ import React from 'react';
 import './Main.scss';
 
 const Main = () => {
-  return <div>Main</div>;
+  return (
+    <div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>
+      <div>Main</div>Main
+    </div>
+  );
 };
 
 export default Main;
